\name{caradat}
\alias{caradat}
\docType{data}
\title{
Locomotor activiy of a Carabid beetle.}

\description{
Locomotor activiy of a blind beetle.
}

\usage{data("caradat")}
\format{
  A data frame with 2014 observations on the following 2 variables.
  \describe{
    \item{Date}{a numeric vector with date and time of day}
    \item{Activity}{a numeric vector of locomotor activity}
  }
}
\details{
activity under DD in a cave observed at UNEVEN intervals.
}

\source{
The data were kindly provided by F. Weber, Münster, Germany. All the experiments were carried out between 1973 and 1980. The dates (but not their order and times) in the file are fictitious.
} 
